Due to a missing break statement, when entering an amex card number the max character length for the field was not being properly set, this has now been capped at 15 digits
