+++ judo-sdk/src/androidTest/java/com/judopay/ui/CardNumberEntryTest.java
@@ -1,0 +1,81 @@
+package com.judopay.ui;
+
+import android.content.Intent;
+import android.support.test.espresso.matcher.ViewMatchers;
+import android.support.test.rule.ActivityTestRule;
+import android.test.suitebuilder.annotation.LargeTest;
+
+import com.judopay.Judo;
+import com.judopay.JudoOptions;
+import com.judopay.PaymentActivity;
+import com.judopay.R;
+import com.judopay.model.Currency;
+
+import org.junit.Rule;
+import org.junit.Test;
+
+import java.util.UUID;
+
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.typeText;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
+
+@LargeTest
+public class CardNumberEntryTest {
+
+    @Rule
+    public ActivityTestRule<PaymentActivity> activityTestRule = new ActivityTestRule<>(PaymentActivity.class, false, false);
+
+    @Test
+    public void shouldHaveAmexCardNumberFormattingWhenAmexCardEntered() {
+        Judo.setAmexEnabled(true);
+
+        activityTestRule.launchActivity(getIntent());
+
+        onView(ViewMatchers.withId(R.id.card_number_edit_text))
+                .perform(typeText("340000432128428"))
+                .check(matches(withText("3400 004321 28428")));
+    }
+
+    @Test
+    public void shouldHaveNormalCardNumberFormattingWhenVisaCardNumberEntered() {
+        activityTestRule.launchActivity(getIntent());
+
+        onView(withId(R.id.card_number_edit_text))
+                .perform(typeText("4976000000003436"))
+                .check(matches(withText("4976 0000 0000 3436")));
+    }
+
+    @Test
+    public void shouldRestrictCardNumberLengthToSixteenDigitsWhenVisa() {
+        activityTestRule.launchActivity(getIntent());
+
+        onView(withId(R.id.card_number_edit_text))
+                .perform(typeText("49760000000034360"))
+                .check(matches(withText("4976 0000 0000 3436")));
+    }
+
+    @Test
+    public void shouldRestrictCardNumberLengthToFifteenDigitsWhenAmex() {
+        activityTestRule.launchActivity(getIntent());
+
+        onView(withId(R.id.card_number_edit_text))
+                .perform(typeText("3400004321284280"))
+                .check(matches(withText("3400 004321 28428")));
+    }
+
+    private Intent getIntent() {
+        Intent intent = new Intent();
+
+        intent.putExtra(Judo.JUDO_OPTIONS, new JudoOptions.Builder()
+                .setJudoId("100407196")
+                .setAmount("0.99")
+                .setCurrency(Currency.GBP)
+                .setConsumerRef(UUID.randomUUID().toString())
+                .build());
+        return intent;
+    }
+
+}
