+++ judo-sdk/src/androidTest/java/com/judopay/ui/PaymentFormTest.java
@@ -38,26 +38,6 @@ public class PaymentFormTest {
     public ActivityTestRule<PaymentActivity> activityTestRule = new ActivityTestRule<>(PaymentActivity.class, false, false);
 
     @Test
-    public void shouldHaveAmexCardNumberFormattingWhenAmexCardEntered() {
-        Judo.setAmexEnabled(true);
-
-        activityTestRule.launchActivity(getIntent());
-
-        onView(ViewMatchers.withId(R.id.card_number_edit_text))
-                .perform(typeText("340000432128428"))
-                .check(matches(withText("3400 004321 28428")));
-    }
-
-    @Test
-    public void shouldHaveNormalCardNumberFormattingWhenVisaCardNumberEntered() {
-        activityTestRule.launchActivity(getIntent());
-
-        onView(withId(R.id.card_number_edit_text))
-                .perform(typeText("4976000000003436"))
-                .check(matches(withText("4976 0000 0000 3436")));
-    }
-
-    @Test
     public void shouldDisplayCidvHintWhenAmexCardNumberEntered() {
         Judo.setAmexEnabled(true);
